fix(logging)!: provide `println!()` when using the `log` facade
