+++ src/ariel-os-log/Cargo.toml
@@ -8,17 +8,52 @@ license.workspace = true
 
 [package.metadata.feature-groups]
 logging-facade.xor = { features = ["defmt", "log"] }
+logging-tranport.xor = { features = ["defmt-rtt", "esp-println", "uart"] }
 
 [dependencies]
+ariel-os-debug = { workspace = true }
+ariel-os-utils = { workspace = true, optional = true }
+const-str = { workspace = true, optional = true }
+critical-section = { workspace = true, optional = true }
 defmt = { workspace = true, optional = true }
+embassy-sync = { workspace = true, optional = true }
 featurecomb = { workspace = true }
 log = { workspace = true, optional = true }
 
+[target.'cfg(context = "esp")'.dependencies]
+esp-println = { workspace = true, optional = true, features = [
+  "auto",
+  "colors",
+  "critical-section",
+] }
+
+[target.'cfg(context = "esp32")'.dependencies]
+esp-println = { workspace = true, optional = true, features = ["esp32"] }
+
+[target.'cfg(context = "esp32c3")'.dependencies]
+esp-println = { workspace = true, optional = true, features = ["esp32c3"] }
+
+[target.'cfg(context = "esp32c6")'.dependencies]
+esp-println = { workspace = true, optional = true, features = ["esp32c6"] }
+
+[target.'cfg(context = "esp32s2")'.dependencies]
+esp-println = { workspace = true, optional = true, features = ["esp32s2"] }
+
+[target.'cfg(context = "esp32s3")'.dependencies]
+esp-println = { workspace = true, optional = true, features = ["esp32s3"] }
+
 [features]
 ## Enables `defmt` as logging facade.
-defmt = ["dep:defmt"]
+defmt = ["dep:defmt", "esp-println?/defmt-espflash"]
 ## Enables `log` as logging facade.
-log = ["dep:log"]
+log = ["dep:ariel-os-utils", "dep:const-str", "dep:critical-section", "dep:log"]
+
+# Logging transports.
+# `defmt-rtt` forces the logging transport to be the debug output, so it has to
+# go through this crate.
+defmt-rtt = ["ariel-os-debug/defmt-rtt"]
+esp-println = ["dep:esp-println"]
+uart = ["dep:embassy-sync"]
 
 _test = []
